Add RTC
When ntpc is run as root, the RTC datetime is also set at the same time. This will ensure that the correct time is maintained by the RTC even after restart. Passing through the If you do not have access to /dev/rtc0, only the RTC setting will be skipped.